Remove clothing/HR models from backend
