--- backend/models/clothing/latest_version.txt
@@ -1,1 +1,0 @@
-20250317_031946
