--- backend/models/clothing/model_meta.json
@@ -1,1 +1,0 @@
-{"version": "20250317_031946", "target": "clothing", "target_columns": ["upr_clo", "lwr_clo"], "model_type": "multi_output_regressor", "combined_r2_score": -0.16349875005439107, "individual_r2_scores": {"upr_clo": -0.7421321528492792, "lwr_clo": 0.4151346527404983}, "timestamp": "2025-03-17T03:19:46.649427", "feature_names": ["temp", "sun", "headwind", "snow", "rain", "fatigued", "hr", "feels"], "best_params": {"max_depth": 10, "min_samples_split": 5, "n_estimators": 200}}
