--- backend/models/hr/latest_version.txt
@@ -1,1 +1,0 @@
-20250317_031945
